pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "proline-cad"
version = "0.1.0"
description = "AI-driven aerospace production line lifecycle planning platform — 3-Agent closed-loop system"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
authors = [
    {name = "ProLine CAD Team"}
]

# ════════════════ 核心依赖 ════════════════
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.30",
    "alembic>=1.13.0",
    "geoalchemy2>=0.14.0",
    "kafka-python>=2.0.2",
    "minio>=7.2.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
# Agent1 — ParseAgent 依赖
parse = [
    "ezdxf>=1.3.0",
    "Rtree>=1.2.0",
    "Shapely>=2.0.4",
    "rdflib>=7.0.0",
]

# Agent2 — ConstraintAgent 依赖
constraint = [
    "z3-solver>=4.12.0",
]

# Agent3 — LayoutAgent 依赖
layout = [
    "numpy>=1.26.0",
    "scipy>=1.13.0",
    "deap>=1.4.1",
    "Rtree>=1.2.0",
    "Shapely>=2.0.4",
]

# LLM 集成
llm = [
    "openai>=1.30.0",
    "tiktoken>=0.7.0",
    "pymilvus>=2.4.0",
]

# 仿真 (Spike-04)
sim = [
    "simpy>=4.1.1",
    "pandas>=2.2.0",
]

# 报告生成 (Spike-10)
report = [
    "reportlab>=4.2.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "Jinja2>=3.1.0",
]

# Dashboard backend (T2 W2 / M1+M2) — observability + queue + auth + uploads
dashboard = [
    "structlog>=24.1.0",
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-api>=1.25.0",
    "opentelemetry-instrumentation-fastapi>=0.46b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.46b0",
    "prometheus-client>=0.20.0",
    "pytest-timeout>=2.3.0",
    # M2 queue + pub/sub
    "arq>=0.26.0",
    "redis>=5.0.0",
    # M2 upload validation
    "python-magic-bin>=0.4.14; platform_system == 'Windows'",
    "python-magic>=0.4.27; platform_system != 'Windows'",
    # M2 JWT auth
    "python-jose[cryptography]>=3.3.0",
]

# 开发工具
dev = [
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
    "httpx>=0.27.0",
]

# 完整安装（所有 Agent + 开发工具 + Dashboard 后端）
all = [
    "proline-cad[parse,constraint,layout,llm,sim,report,dev,dashboard]",
]

[project.urls]
Repository = "https://github.com/hmilyxtl2018/Epic01_ProLine_CAD"
Documentation = "https://github.com/hmilyxtl2018/Epic01_ProLine_CAD/tree/main/PRD"

# ════════════════ Tool 配置 ════════════════

[tool.setuptools.packages.find]
include = ["agents*", "shared*", "app*"]

[tool.ruff]
target-version = "py311"
line-length = 120
src = ["agents", "shared", "spikes"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]
ignore = ["E501"]  # line length handled by formatter

[tool.ruff.lint.isort]
known-first-party = ["agents", "shared"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # 渐进式类型检查
ignore_missing_imports = true

[tool.pytest.ini_options]
# 注意: spikes 的 pytest 配置在 spikes/pytest.ini 中独立维护
# 此处为 agents/ 的测试配置
testpaths = ["agents"]
addopts = "-v --tb=short"
markers = [
    "unit: unit tests",
    "integration: integration tests (require infra)",
    "e2e: end-to-end tests (require all agents)",
]
